(feat)[NavigableEntities] improve management of impact on next entities

+++ src/gws_core/entity_navigator/entity_navigator_service.py
@@ -9,12 +9,10 @@ from gws_core.core.decorator.transaction import transaction
 from gws_core.core.exception.exceptions.bad_request_exception import \
     BadRequestException
 from gws_core.entity_navigator.entity_navigator import (
-    EntityNavigatorExperiment, EntityNavigatorResource)
-from gws_core.entity_navigator.entity_navigator_dto import (
-    DeleteResourceResultDTO, ImpactResult, ProcessResetResultDTO,
-    ResetExperimentResultDTO)
-from gws_core.entity_navigator.entity_navigator_type import (
-    EntityType, NavigableEntity, NavigableEntitySet)
+    EntityNavigator, EntityNavigatorExperiment, EntityNavigatorResource)
+from gws_core.entity_navigator.entity_navigator_deep import NavigableEntitySet
+from gws_core.entity_navigator.entity_navigator_dto import ImpactResult
+from gws_core.entity_navigator.entity_navigator_type import EntityType
 from gws_core.experiment.experiment import Experiment
 from gws_core.experiment.experiment_exception import (
     ResourceUnknownUsedInAnotherExperimentException,
@@ -22,8 +20,10 @@ from gws_core.experiment.experiment_exception import (
 from gws_core.experiment.experiment_service import ExperimentService
 from gws_core.protocol.protocol_model import ProtocolModel
 from gws_core.protocol.protocol_service import ProtocolService
+from gws_core.protocol.protocol_update import ProtocolUpdate
 from gws_core.report.report import Report
 from gws_core.report.report_service import ReportService
+from gws_core.resource.resource_dto import ResourceOrigin
 from gws_core.resource.resource_service import ResourceService
 
 
@@ -37,74 +37,49 @@ class EntityNavigatorService:
 
     @classmethod
     @transaction()
-    def reset_experiment(cls, id_: str, force: bool) -> ResetExperimentResultDTO:
+    def check_impact_for_experiment_reset(cls, id_: str) -> ImpactResult:
         experiment: Experiment = Experiment.get_by_id_and_check(id_)
 
-        impacted_result = cls._get_experiment_reset_impact_result(experiment, force)
-
-        if not impacted_result.success:
-            return ResetExperimentResultDTO(
-                success=False,
-                experiment=experiment.to_dto(),
-                impacted_entities=impacted_result.impacted_entities.get_entity_dict_nav_group()
-            )
-
-        experiment = ExperimentService.reset_experiment(experiment)
-        return ResetExperimentResultDTO(
-            success=True,
-            experiment=experiment.to_dto(),
-        )
+        return cls._calculate_experiment_reset_impact(experiment)
 
     @classmethod
     @transaction()
-    def delete_experiment(cls, experiment_id: str, force: bool) -> ResetExperimentResultDTO:
-        """Delete the experiment
-        """
-        result = cls.reset_experiment(experiment_id, force)
+    def reset_experiment(cls, id_: str) -> Experiment:
+        experiment: Experiment = Experiment.get_by_id_and_check(id_)
 
-        if result.success:
-            ExperimentService.delete_experiment(experiment_id)
+        impact = cls._calculate_experiment_reset_impact(experiment)
+        cls._delete_next_entities(impact.impacted_entities)
 
-        return result
+        return ExperimentService.reset_experiment(experiment)
 
     @classmethod
-    def _get_experiment_reset_impact_result(cls, experiment: Experiment, force: bool) -> ImpactResult:
-        """Method to calculate the impact of the reset of an experiment. If force is True, it will
-        delete the next entities
+    @transaction()
+    def delete_experiment(cls, experiment_id: str) -> None:
+        """Delete the experiment
+        """
+        cls.reset_experiment(experiment_id)
+        cls.reset_experiment(experiment_id)
+
+        ExperimentService.delete_experiment(experiment_id)
+
+    @classmethod
+    def _calculate_experiment_reset_impact(cls, experiment: Experiment) -> ImpactResult:
+        """Method to calculate the impact of the reset of an experiment.
 
         :param experiment: _description_
         :type experiment: Experiment
-        :param force: _description_
-        :type force: bool
         :return: _description_
         :rtype: ImpactResult
         """
+        experiment.check_is_updatable()
         exp_nav = EntityNavigatorExperiment(experiment)
-        next_entities = exp_nav.get_next_entities_recursive([EntityType.EXPERIMENT, EntityType.REPORT])
-
-        # check if there are some experiment that use the result of the experiment
-        if len(next_entities) > 0:
-
-            if not force:
-                return ImpactResult(
-                    success=False,
-                    impacted_entities=next_entities
-                )
-
-            else:
-                cls._delete_next_entities(next_entities)
-
-        return ImpactResult(
-            success=True,
-            impacted_entities=next_entities
-        )
+        return cls._calculate_impact(exp_nav)
 
     ############################################# PROCESS #############################################
 
     @classmethod
     @transaction()
-    def reset_process_of_protocol_id(cls, protocol_id: str, process_instance_name: str,
-                                     force: bool) -> ProcessResetResultDTO:
+    def check_impact_for_process_reset(cls, protocol_id: str, process_instance_name: str) -> ImpactResult:
         """Reset the process of a protocol. To check the impacted entities,
         it check the next entities of the experiment of the protocol
 
@@ -113,19 +88,24 @@ class EntityNavigatorService:
         protocol_model = ProtocolService.get_by_id_and_check(protocol_id)
         protocol_model.check_is_updatable(error_if_finished=False)
 
-        impact_result = cls._get_experiment_reset_impact_result(protocol_model.experiment, force)
+        return cls._calculate_experiment_reset_impact(protocol_model.experiment)
 
-        if not impact_result.success:
-            return ProcessResetResultDTO(
-                success=False,
-                impacted_entities=impact_result.impacted_entities.get_entity_dict_nav_group()
-            )
+    @classmethod
+    @transaction()
+    def reset_process_of_protocol_id(cls, protocol_id: str, process_instance_name: str) -> ProtocolUpdate:
+        """Reset the process of a protocol. To check the impacted entities,
+        it check the next entities of the experiment of the protocol
 
-        protocol_update = ProtocolService.reset_process_of_protocol(protocol_model, process_instance_name)
-        return ProcessResetResultDTO(
-            success=True,
-            protocol_update=protocol_update.to_dto(),
-        )
+        """
+
+        protocol_model = ProtocolService.get_by_id_and_check(protocol_id)
+        protocol_model.check_is_updatable(error_if_finished=False)
+
+        impact_result = cls._calculate_experiment_reset_impact(protocol_model.experiment)
+
+        cls._delete_next_entities(impact_result.impacted_entities)
+
+        return ProtocolService.reset_process_of_protocol(protocol_model, process_instance_name)
 
     @classmethod
     @transaction()
@@ -137,11 +117,13 @@ class EntityNavigatorService:
         protocol_model.check_is_updatable(error_if_finished=False)
 
         # check if there are some experiment that use the result of the experiment
-        reset_result = cls._get_experiment_reset_impact_result(protocol_model.experiment, False)
+        reset_result = cls._calculate_experiment_reset_impact(protocol_model.experiment)
 
         # if yes, raise an error, no force reset for this mode
-        if not reset_result.success:
-            experiments: List[Experiment] = reset_result.impacted_entities.get_entity_by_type(EntityType.EXPERIMENT)
+        if reset_result.has_entities():
+            experiments: List[Experiment] = reset_result.impacted_entities.get_entity_by_type(
+                EntityType.EXPERIMENT)
+
             if len(experiments) > 0:
                 raise ResourceUnknownUsedInAnotherExperimentException(
                     experiments[0].get_short_name(), experiments[0].id)
@@ -162,53 +144,81 @@ class EntityNavigatorService:
 
     @classmethod
     @transaction()
-    def delete_resource(cls, resource_id: str, force: bool) -> DeleteResourceResultDTO:
+    def check_impact_delete_resource(cls, resource_id: str) -> ImpactResult:
+        resource = ResourceService.get_by_id_and_check(resource_id)
+
+        # If the resource was generated by an experiment, reset the experiment
+        if resource.experiment:
+            exp_nav = EntityNavigatorExperiment(resource.experiment)
+            return cls._calculate_impact(exp_nav)
+        else:
+            resource_nav = EntityNavigatorResource(resource)
+            return cls._calculate_impact(resource_nav)
+
+    @classmethod
+    @transaction()
+    def delete_resource(cls, resource_id: str, allow_s3_project_storage: bool = False) -> None:
         """Delete the resource
         """
 
         resource = ResourceService.get_by_id_and_check(resource_id)
-        exp_nav = EntityNavigatorResource(resource)
-        next_entities = exp_nav.get_next_entities_recursive([EntityType.EXPERIMENT, EntityType.REPORT])
 
-        # check if there are some experiment that use the result of the experiment
-        if len(next_entities) > 0:
+        if resource.origin == ResourceOrigin.S3_PROJECT_STORAGE and not allow_s3_project_storage:
+            raise BadRequestException(
+                "Cannot delete the resource because it is an S3 resource")
 
-            if not force:
-                return DeleteResourceResultDTO(
-                    success=False,
-                    resource=resource.to_dto(),
-                    impacted_entities=next_entities.get_entity_dict_nav_group()
-                )
-            else:
-                cls._delete_next_entities(next_entities)
+        # If the resource was generated by an experiment, reset the experiment
+        if resource.experiment:
+            # call delete experiment
+            cls.delete_experiment(resource.experiment.id)
+        else:
+            resource_nav = EntityNavigatorResource(resource)
+            result = cls._calculate_impact(resource_nav)
+            cls._delete_next_entities(result.impacted_entities)
 
-        # Check how to manage s3 resources
-        ResourceService.delete(resource_id)
-        return DeleteResourceResultDTO(
-            success=True,
-            resource=resource.to_dto(),
-        )
+            ResourceService.delete(resource_id)
 
     ############################################# OTHERS #############################################
 
     @classmethod
+    def _calculate_impact(cls, entity_navigator: EntityNavigator) -> ImpactResult:
+        """Method to calculate the impact of the reset of an experiment.
+
+        :param experiment: _description_
+        :type experiment: Experiment
+        :return: _description_
+        :rtype: ImpactResult
+        """
+        next_entities = entity_navigator.get_next_entities_recursive(
+            [EntityType.EXPERIMENT, EntityType.REPORT])
+
+        return ImpactResult(
+            impacted_entities=next_entities
+        )
+
+    @classmethod
     @transaction()
-    def _delete_next_entities(cls, entities: NavigableEntitySet[NavigableEntity]) -> None:
+    def _delete_next_entities(cls, entities: NavigableEntitySet) -> None:
         """Delete the entities
         """
         cls._check_validated_entities(entities, [EntityType.EXPERIMENT, EntityType.REPORT])
 
-        reports = entities.get_entity_by_type(EntityType.REPORT)
+        reports = entities.get_entities_from_deepest_level(EntityType.REPORT)
         for report in reports:
             ReportService.delete(report.id)
 
-        experiments = entities.get_entity_by_type(EntityType.EXPERIMENT)
+        # TODO to improve because this is not perfect if I have 3 experiment.
+        # Exp 1: output --> A
+        # Exp 2: input --> A, output --> B
+        # Exp 3: input --> A, B
+        # In this case the Exp 3 will have the same deep level as the Exp 2
+        experiments = entities.get_entities_from_deepest_level(EntityType.EXPERIMENT)
         for experiment in experiments:
             ExperimentService.delete_experiment(experiment.id)
 
     @classmethod
     def _check_validated_entities(
-            cls, entities: NavigableEntitySet[NavigableEntity],
+            cls, entities: NavigableEntitySet,
             entity_types: List[EntityType]) -> None:
         """Check if there are some validated entities in the set
         """
